New login parameter structure
