+++ LineSDK/LineSDK/Login/LoginManager.swift
@@ -367,53 +367,3 @@ extension LoginManager {
     }
 }
 
-extension LoginManager {
-    /// Represents the language used in web page.
-    public struct WebPageLanguage {
-        public let rawValue: String
-
-        /// Creates a web page language with a given raw string language code value.
-        ///
-        /// - Parameter rawValue: The value represents the language code.
-        public init(rawValue: String) {
-            self.rawValue = rawValue
-        }
-
-        /// The Arabic language.
-        public static let arabic = WebPageLanguage(rawValue: "ar")
-        /// The German language.
-        public static let german = WebPageLanguage(rawValue: "de")
-        /// The English language.
-        public static let english = WebPageLanguage(rawValue: "en")
-        /// The Spanish language.
-        public static let spanish = WebPageLanguage(rawValue: "es")
-        /// The French language.
-        public static let french = WebPageLanguage(rawValue: "fr")
-        /// The Indonesian language.
-        public static let indonesian = WebPageLanguage(rawValue: "id")
-        /// The Italian language.
-        public static let italian = WebPageLanguage(rawValue: "it")
-        /// The Japanese language.
-        public static let japanese = WebPageLanguage(rawValue: "ja")
-        /// The Korean language.
-        public static let korean = WebPageLanguage(rawValue: "ko")
-        /// The Malay language.
-        public static let malay = WebPageLanguage(rawValue: "ms")
-        /// The Brazilian Portuguese language.
-        public static let portugueseBrazilian = WebPageLanguage(rawValue: "pt-BR")
-        /// The European Portuguese language.
-        public static let portugueseEuropean = WebPageLanguage(rawValue: "pt-PT")
-        /// The Russian language.
-        public static let russian = WebPageLanguage(rawValue: "ru")
-        /// The Thai language.
-        public static let thai = WebPageLanguage(rawValue: "th")
-        /// The Turkish language.
-        public static let turkish = WebPageLanguage(rawValue: "tr")
-        /// The Vietnamese language.
-        public static let vietnamese = WebPageLanguage(rawValue: "vi")
-        /// The Simplified Chinese language.
-        public static let chineseSimplified = WebPageLanguage(rawValue: "zh-Hans")
-        /// The Traditional Chinese language.
-        public static let chineseTraditional = WebPageLanguage(rawValue: "zh-Hant")
-    }
-}
