+++ LineSDK/LineSDK/Login/LoginManagerParameters.swift
@@ -1,0 +1,88 @@
+//
+//  LoginManagerParameters.swift
+//
+//  Copyright (c) 2016-present, LINE Corporation. All rights reserved.
+//
+//  You are hereby granted a non-exclusive, worldwide, royalty-free license to use,
+//  copy and distribute this software in source code or binary form for use
+//  in connection with the web services and APIs provided by LINE Corporation.
+//
+//  As with any software that integrates with the LINE Corporation platform, your use of this software
+//  is subject to the LINE Developers Agreement [http://terms2.line.me/LINE_Developers_Agreement].
+//  This copyright notice shall be included in all copies or substantial portions of the software.
+//
+//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
+//  INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+//  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
+//  IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
+//  DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+//  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+//
+
+import Foundation
+
+extension LoginManager {
+
+    public struct Parameters {
+        public var onlyWebLogin: Bool = false
+        public var botPromptStyle: BotPrompt? = nil
+        public var preferredWebPageLanguage: WebPageLanguage? = nil
+    }
+}
+
+extension LoginManager {
+    
+    public enum BotPrompt: String {
+        case normal
+        case aggressive
+    }
+    
+    /// Represents the language used in web page.
+    public struct WebPageLanguage {
+        public let rawValue: String
+
+        /// Creates a web page language with a given raw string language code value.
+        ///
+        /// - Parameter rawValue: The value represents the language code.
+        public init(rawValue: String) {
+            self.rawValue = rawValue
+        }
+
+        /// The Arabic language.
+        public static let arabic = WebPageLanguage(rawValue: "ar")
+        /// The German language.
+        public static let german = WebPageLanguage(rawValue: "de")
+        /// The English language.
+        public static let english = WebPageLanguage(rawValue: "en")
+        /// The Spanish language.
+        public static let spanish = WebPageLanguage(rawValue: "es")
+        /// The French language.
+        public static let french = WebPageLanguage(rawValue: "fr")
+        /// The Indonesian language.
+        public static let indonesian = WebPageLanguage(rawValue: "id")
+        /// The Italian language.
+        public static let italian = WebPageLanguage(rawValue: "it")
+        /// The Japanese language.
+        public static let japanese = WebPageLanguage(rawValue: "ja")
+        /// The Korean language.
+        public static let korean = WebPageLanguage(rawValue: "ko")
+        /// The Malay language.
+        public static let malay = WebPageLanguage(rawValue: "ms")
+        /// The Brazilian Portuguese language.
+        public static let portugueseBrazilian = WebPageLanguage(rawValue: "pt-BR")
+        /// The European Portuguese language.
+        public static let portugueseEuropean = WebPageLanguage(rawValue: "pt-PT")
+        /// The Russian language.
+        public static let russian = WebPageLanguage(rawValue: "ru")
+        /// The Thai language.
+        public static let thai = WebPageLanguage(rawValue: "th")
+        /// The Turkish language.
+        public static let turkish = WebPageLanguage(rawValue: "tr")
+        /// The Vietnamese language.
+        public static let vietnamese = WebPageLanguage(rawValue: "vi")
+        /// The Simplified Chinese language.
+        public static let chineseSimplified = WebPageLanguage(rawValue: "zh-Hans")
+        /// The Traditional Chinese language.
+        public static let chineseTraditional = WebPageLanguage(rawValue: "zh-Hant")
+    }
+}
